Add option to limit the number of rows in table
Sometimes, the table is rendered externally or with few outside parameters and
turns out to be too long to be displayed properly. It requires some slice
modifications in order to truncate the list and to introduce an ellipsis to
indicate that the table row count was reduced.

Add table option to tell the table renderer a limit on how many rows should be
added to the final result. An ellipsis is added to indicate that the number of
rows was truncated.