Fix weird table row limit behaviour
If a table row limit is specified that is greater than the actual size
of the table, an error is created. This makes sense from a programming
point of view, however it is counter intuitive when using the feature as
most developers would assume that if the row limit is greater than the
row count, only the available content is displayed.

Remove `RowLimitExceedsTableSize` error struct and code.

Change row limit condition to only set a row limit if it is in the
bounds of the table rows.

Fix test case to not test for an out of bounds error, but that the table
is printed in its entirety even though a high row limit is set.