Fix indention problem with lists in lists
In `dyff`, an issue was reported that showed the wrong usage of the
indention prefix in lists in lists:
```
cluster-1
  - one list entry removed:     + two list entries added:
    - │ - x1                      - │ - x1
    │ - x5                        │ - x5
                                  │ - x10
                                  - x999
```

The correct style would be like this:
```
cluster-1
  - one list entry removed:     + two list entries added:
    - - x1                        - - x1
    │ - x5                        │ - x5
                                  │ - x10
                                  - x999
```

Add two additional checks in the sequence node code to correctly use the
indention prefix only in list entries other than the first one. An
exception of that rule is in case `skipIndentOnFirstLine` is `false`.