Add neat output function for struct
In case the input object was a generic type struct, the switch for the
neat output did not pick it up and used the scalar function to render
the struct, which resulted in an output without any highlights.

Add `reflect` based check to look for generic struct objects.

Add function to render struct objects.

Add test case to verify that a generic struct is rendered correctly.